backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
PyGithub==2.3.0
requests==2.32.3
SQLAlchemy==2.0.34
flake8==7.1.1
black==24.10.0
